gpu: Fix invalid parameter type of glfw framebuffer callback function

+++ gpu/src/dawn/hello_triangle.zig
@@ -116,7 +116,7 @@ pub fn main() !void {
     };
     setup.window.setUserPointer(CallbackPayload, &.{ .swap_chain = swap_chain, .swap_chain_format = swap_chain_format });
     setup.window.setFramebufferSizeCallback((struct {
-        fn callback(window: glfw.Window, width: isize, height: isize) void {
+        fn callback(window: glfw.Window, width: u32, height: u32) void {
             const pl = window.getUserPointer(*CallbackPayload);
             c.wgpuSwapChainConfigure(pl.?.swap_chain, pl.?.swap_chain_format, c.WGPUTextureUsage_RenderAttachment, @intCast(u32, width), @intCast(u32, height));
         }
